+++ src/Ampersand/ADL1/Disambiguate.hs
@@ -129,6 +129,10 @@ instance Disambiguatable P_ViewD where
                         , constraints
                         )
    where constraints = Cnstr [MustBe (pCpt2aCpt cptMap c)] []
+instance Disambiguatable P_Enforce where
+  disambInfo cptMap (P_Enforce o a op b) env1 = ( P_Enforce o a' op b', propagateConstraints envA envB )
+   where (a', envA) = ( (a, env1), Cnstr (getDConcepts source (snd a)) (getDConcepts target (snd a)))
+         (b', envB) = disambInfo cptMap b (propagateConstraints env1 envA)
 
 instance Disambiguatable P_ViewSegment where
   disambInfo cptMap (P_ViewSegment a b c) i = (P_ViewSegment a b c', r)